requirements.txt
PySide6
yt_dlp
ffmpeg
torch
openai-whisper
openai
# pandas
# numpy
# librosa
# soundfile
# tensorflow
# tensorflow-hub